Use patch_fetch_from_pypi in check_unused test
